move visual-related types into new common::visual
Previously, visual-related types were strewn throughout `connection` and `common`.
