+++ src/x11.rs
@@ -13,4 +13,4 @@
 pub mod error;
 pub mod event;
 // pub mod reply;
-// pub mod request;
+pub mod request;
